+++ source/units/ld_comparation.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include <cmath>
 
-constexpr bool equal(long double value_1, long double value_2) {
-    constexpr long double EPS = 1e-6;
+constexpr bool equal(double value_1, double value_2) {
+    constexpr double EPS = 1e-6;
     return std::abs(value_1 - value_2) < EPS;
 }
