+++ source/units/speed.hpp
@@ -11,7 +11,7 @@ public:
     constexpr explicit Speed() : m_value_bit_per_ns(0) {}
     
     // Attention: value given in TSizeBase per TTimeBase units!
-    constexpr explicit Speed(long double value)
+    constexpr explicit Speed(double value)
         : m_value_bit_per_ns(value * TSizeBase::to_bit_multiplier /
                              TTimeBase::to_nanoseconds_multiplier) {}
 
@@ -19,14 +19,12 @@ public:
     constexpr Speed(Speed<USizeBase, UTimeBase> speed)
         : m_value_bit_per_ns(speed.value_bit_per_ns()) {}
 
-    constexpr long double value() const {
+    constexpr double value() const {
         return m_value_bit_per_ns / TSizeBase::to_bit_multiplier *
                TTimeBase::to_nanoseconds_multiplier;
     }
 
-    constexpr long double value_bit_per_ns() const {
-        return m_value_bit_per_ns;
-    }
+    constexpr double value_bit_per_ns() const { return m_value_bit_per_ns; }
 
     constexpr ThisSpeed operator+(ThisSpeed speed) const {
         return Speed<Bit, Nanosecond>(m_value_bit_per_ns +
@@ -38,11 +36,11 @@ public:
                                       speed.m_value_bit_per_ns);
     }
 
-    constexpr ThisSpeed operator*(long double mult) const {
+    constexpr ThisSpeed operator*(double mult) const {
         return Speed<Bit, Nanosecond>(m_value_bit_per_ns * mult);
     }
 
-    constexpr long double operator/(ThisSpeed speed) const {
+    constexpr double operator/(ThisSpeed speed) const {
         return m_value_bit_per_ns / speed.value_bit_per_ns();
     }
 
@@ -53,9 +51,9 @@ public:
         m_value_bit_per_ns -= speed.m_value_bit_per_ns;
     }
 
-    constexpr void operator*=(long double mult) { m_value_bit_per_ns *= mult; }
+    constexpr void operator*=(double mult) { m_value_bit_per_ns *= mult; }
 
-    constexpr void operator/=(long double mult) { m_value_bit_per_ns /= mult; }
+    constexpr void operator/=(double mult) { m_value_bit_per_ns /= mult; }
 
     bool constexpr operator<(ThisSpeed speed) const {
         return m_value_bit_per_ns < speed.m_value_bit_per_ns;
@@ -74,7 +72,7 @@ public:
     }
 
 private:
-    long double m_value_bit_per_ns;  // value in bit per nanosecond
+    double m_value_bit_per_ns;  // value in bit per nanosecond
 };
 
 template <IsSizeBase TSizeBase, IsTimeBase TTimeBase>
